Fix project structure
